Bug Fix: Open form for only signed in users
